Time: 2 ms (81.91%), Space: 42.9 MB (33.13%) - LeetHub

+++ 0171-excel-sheet-column-number/0171-excel-sheet-column-number.java
@@ -1,8 +1,8 @@
 class Solution {
     public int titleToNumber(String columnTitle) {
-        long multiplyFactor = 1;
+        int multiplyFactor = 1;
         int n = columnTitle.length();
-        long ans = 0;
+        int ans = 0;
         for(int i = n-1; i>=0; --i){
             int current_Value = columnTitle.charAt(i) - 'A' + 1;
             ans = ans + multiplyFactor*current_Value;
